Refactored - moved cookie generation from the BaseController to the CookieService. Fixed unit test.
